dagql: support dynamic purity
Currently impurity is static at the schema level: if a field is marked
impure, it will never be cached, and IDs coming out of it will be
tainted.

This quickly improved insufficient: we want some APIs to determine
whether they're pure at runtime, based on their inputs.

For example, we want this call:

    cacheVolume(key: "foo")

to return a namespaced volume, expanding implicitly to:

    cacheVolume(key: "foo", namespace: "modID").

With this change, we can just have cacheVolume call dagql.Taint(ctx) if
no namespace is provided.

Signed-off-by: Alex Suraci <alex@dagger.io>