Fix broken terminal tests and broken Directory.terminal feature (#8952)
* tests: fix terminal tests not asserting as expected

Our terminal tests were not actually working how we thought they were.
Many of them were not waiting for prompts and instead just matching on
strings right away, which ended up including all the progress output
before the terminal was opened.

This broke a few of them since the string they were expecting happened
to also show up in the progress output, which isn't what we actually
wanted to assert on.

This updates the tests to correctly wait for prompts (doing some
terminal escape sequence handling when needed).

This commit leaves the `TestDaggerTerminal/directory` test case failing
because that feature was actually totally broken due to a separate issue
(which went unnoticed until now because the test case was broken).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: fix Directory.terminal and overlay mount handling

The previous commit fixed assertions in our terminal test cases, which
revealed that the Directory.terminal feature was totally broken.

The problem was in the executor and it's handling of mounts that were of
type overlay. It was looking at the `Source` field of the mount assuming
it was a bind and that the field pointed to the source of the bind.

However, for non-bind mounts the `Source` is just the type of mount.

The reason this only broke the Directory.terminal case so far was that
Directory.terminal is the only place in our code where we create
read-only mounts. This triggers a different codepath in buildkit's cache
logic that results in actual overlay mounts being provided to the
executor; in other read-write mount cases buildkit has some logic that
creates a "shared" overlay mount which is just bind-mounted to other
places (to avoid conflicts around multiple overlay mounts with the same
upperdir).

This commit just fixes handling of that `Source` field to interpret it
correctly based on whether it's a bind mount or not.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>