bump OTel to 1.32.0, semconv to 1.26.0
This is the first time bumping since adding the replace pinning to
module go.mods, and it took a bit of a rework to deal with go.work.

In particular, any modules that rely on go.work need to additionally
have a replace rule installed to go.work, otherwise you'll get an error
as soon as you bump one module's individual replace rules.

So now the go.mod and go.work replacing happens via PostCommands, which
also avoids the issue of modfile.ParseLax not parsing replace rules.
Also, now we go mod tidy before `go work use .` which is important for
bumping the Go version that ends up in go.work.

Signed-off-by: Alex Suraci <alex@dagger.io>