ci: Merge smaller `Engine & CLI split` jobs into `Engine & CLI`
We introduced an alternative workflow which splits the big test suites
into smaller ones that have appropriately-sized CI runners. Context:
- https://github.com/dagger/dagger/pull/7863
- https://github.com/dagger/dagger/pull/8212

Now that we have been running both workflows in parallel for 4 months
and have a long-term understanding of their behaviour, we are going to
**temporarily** pause the less reliable jobs and save on infra
costs. For more details on the two workflows see this original issue
description which has been kept up-to-date all these months:
- https://github.com/dagger/dagger/issues/8184

As a reminder, we are doing this change so that we:
- mark a new milestone in this saga which started 6 months ago, in
  May 2024: https://github.com/dagger/dagger/issues/7339
- make progress towards `All our jobs succeed 90% of the time and
  finish within 10 minutes` (this change effectively makes our longest
  job be ~11mins)
- make all contributors wait less for all PR checks to be green
  (including re-running jobs which fail intermittently)
- save on CI infra costs which have doubled since September and are on
  track to reach `$10k` just for November

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>