ci(alt nightly): Disable our heaviest pipelines on alternative runners & run twice daily (#8999)
* ci(alt nightly): Disable our heaviest pipelines on alternative runners

They have been failing days in a row since they require at least 2x the
CPU resources that are made available by default. We could bump the
timeout, but that will only result in higher infra costs and not provide
much value - we already know about the limitations & current behaviour.

We are also only keeping pipelines for a subset of SDKs. The goal is to
have decent coverage, not be exhaustive.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci(alt nightly): Change the workflow cron trigger

Run it twice so that we test the caching efficiency.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>