ci: make engine generate return a minimal diff
Without this, we'd actually keep all the contents like `docs` entirely
as is - when we later composed all the files together, we'd actually
overwrite the newly generated files. This meant that `generate` didn't
actually regenerate any files on an API schema change.

So we can just take the diff, and now we can compose these all together!

Signed-off-by: Justin Chadwell <me@jedevc.com>