engine: fix goroutine leaks + deadlock in terminal (#9013)
* engine: slightly optimize terminal reading by reusing buffers

Rather than allocating a new byte slice for every read from
stdout/stderr sent to clients, we can re-use the same byte slice.

This reduces the amount of allocations the engine does. Not a
world-changing optimization but I have seen similar fixes in other
contexts make a noticeable difference in CPU usage since the Go runtime
has a lot less work to do.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* engine: fix goroutine leaks + deadlock in terminal

Before this, every interactive debug terminal container would leak
multiple goroutines due to various io pipes not getting closed. In one
particular codepath, the read-side of the stdout io.Pipe would be left
open while the other side was in the middle of a write, which caused the
executor code to block forever and never return. This manifested as an
occasional flake in the TestModule/TestDaggerTerminal/on_failure test
case.

Additionally, the part of that test case which meant to assert on
handling of --interactive-command values which didn't exist was not
checking error values. It was hitting an error due to the fact that it
didn't have a TTY setup rather than the error around the command not
existing. That test case is fixed now to setup a TTY and verify the
error message is the expected one.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>