fix: allow tui to open /dev/tty as fallback
If it's available, we should try to fallback to `/dev/tty` as a fallback.
This logic is *borrowed* from bubbletea, however, we can't just use
`WithInputTTY`, since this requires that `/dev/tty` exists - we want to
continue on, even if it doesn't.

This allows us to neatly handle the `echo "" | dagger shell` case -
stdin is not a TUI, but we can still usually find one at `/dev/tty`, so
we should try and do that.

Note: we also do a little refactor of `findTTYs` to directly return
`io.Reader`/`io.Writer`s - this avoids the horror of typed nils, which
can result in passing a `nil` `os.File` into bubbletea (which is *does
not like*).

Signed-off-by: Justin Chadwell <me@jedevc.com>