fix: insert `Module.withDependencies` in correct order
Previously, calling `dagger install` with no name would bizarrely insert
it into the beginning of the `dagger.json` `dependencies` field. Then
calling `dagger develop` would fix it.

This is what was happening:
- We call `updateDeps` to update our dependencies.
- We were calling `ModuleSource.dependencies` (which merges
  `withDependencies` nicely) and getting all ordered dependencies.
- However, no name was specified in the `withDependencies` call, so the
  name was actually the empty string - so was appearing first in the
  lexigraphic sort.
- Then, later in `updateDeps`, we fill in the missing name - this is
  correct, but is now too late for the sort, so we have the wrong order.

The solution is fairly simply - `dependencies` should fill in missing
names if possible.

We *could* have added `updateDeps` as an alternative, but this is
actually not great, since this means that `dependencies` would still be
returning `dependencies` that have no name - which is a potential
footgun for the next person coming along.

Signed-off-by: Justin Chadwell <me@jedevc.com>