chore: consolidate shell command Run and RunState
We don't need two separate methods for this - instead, we can validate
it using a similar mechanism for how we do the args.

This will allow us to more easily determine if a builtin supports a
context input, so we can choose whether it should be part of
auto-complete in a certain context.

Signed-off-by: Justin Chadwell <me@jedevc.com>