engine: pick fix for releasing unreferenced cache keys on gc
This was fixed in upstream but needs a mix of updating our
dagger/buildkit fork and some of our own in-repo code.

It results in more cache entries that are no longer referenced after a
GC being pruned (previously they were only released with an explicit
prune API call or on engine reboot).

Signed-off-by: Erik Sipsma <erik@sipsma.dev>