fix(engine): secrets created inside module constructor fail on repeated accesses (#8948)
* successfully reproducing in test

Signed-off-by: Connor Braa <connor@dagger.io>

* fix: ensure we transfer secrets from correct client for constructor

Our previous clever fix which involved setting all function evaluations
as impure (and relying entirely on buildkit's caching) is very good, and
allows us to have a little bit smarter secret retrieval logic.

However! We were actually weirdly resetting the constructor functions as
pure. This really isn't the right thing to do, I missed this case when
working on this originally - we can just remove this code, and rely on
the call to `FieldSpec` to do the right thing (tm).

Hopefully this will just all go away when we remove `SetSecret`.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* error message improvements

Signed-off-by: Connor Braa <connor@dagger.io>

---------

Signed-off-by: Connor Braa <connor@dagger.io>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>