fix(vcs): Make `RepoRootForImportDynamic` trims trailing `../` (#9094)
CI was recently broken due to a change on GitLab's end, when parsing a ref.
In order to isolate the root of the repo to the subdir, we rely on Go's logic:
1. Statically analyze the ref
2. If no match, query the git provider with `ref?go-get=1`

It turns out that GitLab recently introduced a change that normalizes refs with `../`, when querying
refs such as `gitlab.com/foo/bar/baz/../..?go-get=1`, it now becomes: `gitlab.com/foo/?go-get=1`

I suspect this change: https://gitlab.com/gitlab-org/gitlab-foss/-/blame/master/lib/gitlab/middleware/go.rb#L57, introduced 5 months ago, but that could have just been deployed last week.

Several options were considerable, but it seems acceptable to directly remove them from the ref itself, when being the trailing elements

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>
Co-authored-by: Guillaume de Rouville <guillaume@dagger.io>