dagui: misc tweaks for frontend optimizing (#9042)
* simplify Simplify

rather than finding the 'smallest' call that produced an output, which
is slow and fraught with peril, just take the call from the earliest
span we saw for it. which also seems much more likely to be the most
recognizable form.

Signed-off-by: Alex Suraci <alex@dagger.io>

* dagui: add DB.RemainingSnapshots

this can be used by the backend to sync all the unsent snapshots over
the client so it can go away.

Signed-off-by: Alex Suraci <alex@dagger.io>

* respect effect ID cause/effect correlation

not too tough to deal with, seems worth it so Cloud can view pre-v0.15
traces

Signed-off-by: Alex Suraci <alex@dagger.io>

* dagui: never consider a Passthrough span as Hidden

Passthrough takes precedence over everything. Its purpose is to
literally drop out of the plane of existence and jump straight to the
children; it isn't meant as a shortcut to "shroud" the children.

Signed-off-by: Alex Suraci <alex@dagger.io>

* dagui: mark calls final if followed by non-call

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert "dagui: never consider a Passthrough span as Hidden"

this was motivated by v2 compatibility, but it is way too noisy

This reverts commit dd6a0ef6669ce7df1946d1cb3754044f45e0387b.

Signed-off-by: Alex Suraci <alex@dagger.io>

* encapsulate all buildkit non-vertex spans

* "resolving ..." _generally_ isn't worth calling out unless it fails
* now we don't need to check for specific span names

Signed-off-by: Alex Suraci <alex@dagger.io>

* add Activity.EndTime

Signed-off-by: Alex Suraci <alex@dagger.io>

* cleaner tracking of span versions

* don't mutate when snapshotting, mutate when propagating to parents
* increment parents when new child arrives so ChildCount gets bumped

with these changes we shouldn't need to do the awkward recursive version
summing on the client side anymore

tangentially related, count hidden children too, to eliminate weird
dependency on verbosity level; we should indicate that there are hidden
spans instead

Signed-off-by: Alex Suraci <alex@dagger.io>

* only consider the earliest cause for activity

for compatibility with pre-v0.15 which added effects to all spans,
rather than only ones that introduced the effect

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix frontend losing running status

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix frontend losing passthrough status for sync

Signed-off-by: Alex Suraci <alex@dagger.io>

* dont consider span cached if it has children/logs

a bit misleading - if there are children or logs, something obviously
ran

Signed-off-by: Alex Suraci <alex@dagger.io>

* add IsCanceled helper which detects root ending

we don't want to represent a span as 'running' if the root span clearly
ended

Signed-off-by: Alex Suraci <alex@dagger.io>

* remove Activity.EarliestRunningID

this was a precursor to AllRunning - not needed anymore

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix handling of simplification cycles

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix potential missed change marker

if we updated the earliest running timestamp, we should for sure mark
that as changed, even if the final interval ends up subsumed by existing
ones

Signed-off-by: Alex Suraci <alex@dagger.io>

* rename (Links -> Effects)

Signed-off-by: Alex Suraci <alex@dagger.io>

* fix losing version in ImportSnapshots

this only really affects viewing a running trace, which is much harder
to test, hence the telemetry suite not covering it.

tl;dr without this the versions in the web UI would not increment,
leaving things stuck in a 'running' state

Signed-off-by: Alex Suraci <alex@dagger.io>

* be extra careful about remotely calculated states

this is mostly just chasing ghosts in the machine - but it felt extra
murky having some of these getters overloaded by the very field that
they're used to set, since that meant they would never transition from
that state (e.g. Pending true => false)

Signed-off-by: Alex Suraci <alex@dagger.io>

* be more liberal with causal spans

if we only look at links when links are present, services are left
pending

still not 100% sure about this, to be honest, but the behavior "feels
good", judging by:

    dagger-dev call test specific --run CLI/Install/GitLab/git/unpinned

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert "fix handling of simplification cycles"

This reverts commit 12f11898349edd9654f03b12ea1b074a964f80f4.

Signed-off-by: Alex Suraci <alex@dagger.io>

* Revert "simplify Simplify"

This reverts commit c49ed0451a265bd6f192bf7570510662739ccf58.

Signed-off-by: Alex Suraci <alex@dagger.io>

* chore: rename remote to final

Signed-off-by: Justin Chadwell <me@jedevc.com>

* tests: fix broken viztest

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>
Signed-off-by: Justin Chadwell <me@jedevc.com>
Co-authored-by: Justin Chadwell <me@jedevc.com>