ci: Add bump-dagger-version function to daggerverse (#8947)
Upgrades Dagger version for Daggerverse in:
- GitHub workflows
- Daggerverse modules
- Daggerverse app
- Infra config

It also deploys a Daggerverse preview environment with a development
version of Dagger (this implies the version that we are upgrading to).

This now runs automatically part of the publish workflow. When no
`--from` is specified, it looks inside the `.changes` directory and
picks the one but last version (n-1).

Updated the publish workflow runs-on to only use our own runners. This
workflow is never meant to run in forks.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>