Add support for custom ID digests to enable content hashing
This enables implementations of API calls to set a custom digest on
returned IDs replacing the default digest calculated by the "call chain
digest".

Currently, this is used to enable content-based caching of certain calls
that previously relied on the separate `blob` API. The `blob` API is
thus removed as part of this change too.

The intention is to setup more use cases for this, including customizing
digests of different ModuleSource implementations and reducing cache
invalidation of SDK calls.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>