Update TestCustomSDK to use correct arg type
The type of introspectionJson changed from string to *File a long time
ago, but for some reason dagql never error'd out despite the type
mismatch (perhaps something to do with coercing an ID to a string?).

The previous commit accidentally fixed that validation, so this test
started correctly failing now and needs to be corrected.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>