helm: Enable deploying multiple Engines per node
This adds a new property - `kind` - which defaults to `DaemonSet`. It
means that nothing will change for existing users.

For users that want to be able to deploy **multiple** Engines on a
single node, they just need to set `kind: StatefulSet`, and the chart
will take care of the rest.

This opens the door to experimenting with runtimes such as
firecracker-containerd, gVisor, etc. for putting a sandbox around the
Engine - should help address some of the `CAP_SYS_ADMIN` concerns.

Paired with Yannik Dällenbach & Christoph Raaflaub

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>