helm: Enable deploying multiple Engines per node (#9249)
* helm: Enable deploying multiple Engines per node

This adds a new property - `kind` - which defaults to `DaemonSet`. It
means that nothing will change for existing users: one Engine per K8s
node.

For users that want to deploy **multiple** Engines per K8s node, set
`kind: StatefulSet`, and the chart will take care of the rest.

This opens the door to experimenting with runtimes such as
firecracker-containerd, gVisor, etc. for putting a sandbox around the
Engine - should help address some of the `CAP_SYS_ADMIN` concerns.

Paired with Yannik Dällenbach & Christoph Raaflaub from puzzle.ch

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Test installing a second Engine

This validates that multiple Engines can coexist on the same K8s node,
in the same namespace. We are also checking that the pod owner is
correct: either DaemonSet or StatefulSet.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Add changelog fragment

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Bump test-everything-else to 32CPUs

It's been failing more frequently since:
- https://github.com/dagger/dagger/pull/9195

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Set replicas to 1 explicitly & explain why in a comment above

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>