ci: fix permissions error on pull request comment
If the PR comes from a fork, then we cannot get a write token for
pull-requests, so we can't write a comment to the PR.

To be able to get access to this, we need to instead trigger on
`pull_request_target`. This uses the job definition from `main` instead.

Signed-off-by: Justin Chadwell <me@jedevc.com>