handle & inside javadoc
Javadoc comments are wrapped between HTML `<p>` tags. In this case `&` is
not a valid entity and generates the following error:

    error: bad HTML entity
      * <p><br/>Set chart & app version<br/></p>

This happen when you generate the documentation using `mvnw package -Pjavadoc`

This commit escapes the `&` char and also call the
`Helpers.escapeJavadoc` method for type and arg descriptions in addition to the
already done field one.

Signed-off-by: Yves Brissaud <gh@lgtd.io>