switch to custom init that also handles nested exec sessions (#9262)
This change replaces dumb-init with a custom one written in go that also
serves session attachables for nested execs.

One motivation is to avoid a few minor headaches around cross-compiling
C code (which dumb-init) is written in. Cross-compiling the new go
implementation is trivial.

More than that, this allows transplants session attachables from the
executor to this init process, which greatly reduces the hardship
involved in adding new session attachables.

Rather than having to figure out how to get each attachable to work both
in the CLI and in the executor (which involves complicated entering into
the container namespaces), we can just have a single implementation of
those attachables that works everywhere.

This doesn't go all the way back to the previous shim implementation
(and it's associated headaches), only the attachables are running in the
container again. We also avoid needing to do more "pass data via
internal env var" by instead giving the init one side of a socketpair,
which it uses to connect attachables rather than having to do a big
dance to establish the connection itself using tons of client-specific
metadata.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>