fix: ensure container is cached before starting terminal service
This ensures that we avoid immediately flashing the terminal message,
and not showing any progress messages while we're building the service
container on demand.

Ideally, the TUI would handle this by default - but currently showing the
terminal just entirely suspends all progress output. This needs fixing
*at some point*, but it's worth fixing with this hack now (it doesn't
have any extra cost to do this).

Signed-off-by: Justin Chadwell <me@jedevc.com>