ci: avoid use of errgroup.WithContext in checks
This causes some very hard-to-read CI failures when we use this - this
means that instead of waiting for *everything* to finish, we bail
quickly on the first error. E.g. when linting, we error out on the first
lint failure, instead of waiting for all the checks to pass or fail.

Signed-off-by: Justin Chadwell <me@jedevc.com>