fix: Replace custom module in PhpSdkDev
Thanks to the `:expect` flag on calls to `withExec`
The custom-module _always-exec_ used in the PhpSdkDev is now obsolete.

This PR removes the depedency on the custom module _always-exec_.

Also throw an error for anything that is **not** what we expect,
This is different from prior behaviour which only threw for errors we
expected. (exit code 3)

Signed-off-by: John Charman <john.charman@imhotek.net>