fix excessive memory usage by still using LLB blob source for schema
LLB (un)marshalling happens a ton in buildkit and when a file op takes
up lots of space it consumes a ton of memory.

Unfortunately at the moment the best fix is to still use the LLB blob
source for schemas (since they are large files). The `blob` API in
dagger can still disappear thankfully, only the internal LLB source
still needs to exist.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>