fix: ensure trailing slashes in WithFiles doesn't change behavior (#9457)
The code that propagates operations to a container's underlying
directory wasn't quite correct. It was incorrectly making the assumption
that combining `filepath.Dir` + `filepath.Base` gives you the original path.

However, when passing a path with a trailing slash to `filepath.Dir`,
you actually get the entire input - instead of going up a level. So,
instead, we should avoid this, and instead `filepath.Clean` the input so
that we ensure we never have a trailing slash. We can also make these
methods consistent, and use `filepath.Split` as well (while we're in the
area, this doesn't actually have a behavior change).

Signed-off-by: Justin Chadwell <me@jedevc.com>