sdk/java: implement constructors
Allow to define a constructor on classes for an module object.

- a default, empty, constructor calling `super` is still required
- only the first non empty constructor will be registered

This constructor can have as the first parameter a
`io.dagger.client.Client` arg.
In this case the Dagger client will be injected. Please be
sure to then call super with the client:

    @Object
    public class MyModule extends AbstrctModule {
      // to allow creation of empty object
      // mandatory
      public MyModule() {
        super();
      }

      // this constructor will be registered with
      // a string parameter and the dagger client
      // will automatically be injected at creation
      // so it can be used in the constructor body
      public MyModule(Client dag, String stringArg) {
        super(dag);
	// ...
      }
    }

You do _not_ require to have the client as first parameter.
If your constructor doesn't requires it, it can be omitted.

You can use annotations on constructors' arguments as in
functions. For instance:

    public String version;

    public MyModule(@Default("1.2.3") String version) {
      this.version = version;
    }

On a technical point, the generated code to run module functions
has changed. Before, the code was using reflexivity. But this is
not necessary. This code is not generic, it's a code generated
for a specific module source code. So the reflexivity has been
replaced by direct calls. This simplifies both the code generation
part and the generated code itself.

Signed-off-by: Yves Brissaud <gh@lgtd.io>