Improve releasing during v0.15.3 (#9474)
* ci: bump workers to v0.15.3

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: bump internal tooling to v0.15.3

Signed-off-by: Justin Chadwell <me@jedevc.com>

* ci: fix component github releasing

This had previously broken! The tag + target arguments were getting
confused, and the docs *used* to be the wrong way round.

So this now clarifies - src is the original tag, target is the new tag.
This is now the right way round, and it should work.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* ci: actually print the nice release report

And error out if we detect that it sadly errored :(

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: ensure helm chart is published

Signed-off-by: Justin Chadwell <me@jedevc.com>

* Fix a bunch of outdated dagger.json files

Running `dagger develop` in all those dirs fixed it.

This makes me wonder whether we should automate this and run `dagger
develop` in every dir with a `dagger.json` after a new Dagger release
ships.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Bump Dagger version used in alternative CI Runners

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* fix: avoid unused arg in github release

Signed-off-by: Justin Chadwell <me@jedevc.com>

* chore: fix test-everything-else to use 32 cores again

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>
Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Co-authored-by: Gerhard Lazu <gerhard@dagger.io>