fix sourcemap tests
This was extremely subtle to figure out, but previously our sourcemap
tests used modules that had different context dirs due to no .git being
present (and thus each module's context dir being set to its source
root).

The logic for sourcemaps had an implicit dependency on the fact that
even though the context dirs were different in this case, it just so
happened that they all got loaded together and thus the layout of the
dirs seemed like they were sharing a context dir.

The changes/simplifications to the loading are crucial so we can't just
go back to loading everything together.

For now, source maps will only be accurate when the module dependencies
share a context. This is likely 99.9% of the cases and extremely trivial
to fix for users that ever hit this situation.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>