fix seemingly pre-broken TestContextDirect/as_module
I may easily be missing something but it doesn't seem like this test
should have ever passed.

It provided a directory as an arg and called AsModule on that, which
should have resulted in the given directory being the context dir for
the module.

However it then tried to reference subpaths of that module as though it
was using a different context.

I can't see how this should have been expected to pass and seems
possible there was a previous bug that now caused it to correctly fail.

The adjustments to make it work as expected are trivial, just some
tweaks to paths so the context dir is setup as expected.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>