ci: fix failing helm job with version incompat (#9622)
Due to details in our version comparisons, the recent min version bump
to v0.16 has caused our helm job to start breaking.

This is because the helm job starts an image from *main*. This is not
ideal, and the versions between these are kind of mutually incompatible.
We might want to loosen these at some point a bit, but I don't really
have the time to reason through the implications of that *right* now -
so this should quickly unblock that job from passing.

Signed-off-by: Justin Chadwell <me@jedevc.com>