feat: introduce Anthropic support to `llm` type (#9656)
* feat: add support to other llm clients

Current codebase is currently very tied to openai library, which works for most of models as they are retrocompatible.
However, Anthropic is not.

This gives us the opportunity to implement the model-agnostic way of sending queries

This is a hard rebase from https://github.com/shykes/dagger/pull/297, implemented by samalba

Porting the implementation led to a mismatch in the sendQuery function, introducing a segfault. This fixes it

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* feat: add anthropic support

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

* chore: update dependencies

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>

---------

Signed-off-by: Guillaume de Rouville <guillaume@dagger.io>
Co-authored-by: Sam Alba <216487+samalba@users.noreply.github.com>