sdk/java: replace `@Nullable` by `Optional<?>` arguments (#9605)
* always call dagger for introspection

As the introspection json files have been removed, always call dagger
to get the introspection

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* extract parse parameter function

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* extract code .withFunction

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* extract function invocation code

Do not invoke using reflection but directly by calling the function.
That's the benefit of the code generation instead of reflexivity.

This also allows to deal with param types better because there's no
need to find the right method signature which is easier in preparation
for parameterized types like Optional<String> and others.

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* replace @Nullable by Optional<>

Before:

    @Function
    public String echo(@Nullable String arg) {
      if (arg == null) {
        arg = "dagger!";
      }
      return arg;
    }

After:

    @Function
    public String echo(Optional<String> arg) {
      return arg.orElse("dagger!");
    }

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* add default value for short, byte and char

Signed-off-by: Yves Brissaud <gh@lgtd.io>

---------

Signed-off-by: Yves Brissaud <gh@lgtd.io>