error out in dagger query/listen if dagger.json totally invalid
The fallback behavior of these commands when failing to load a module
meant that even if a dagger.json was found but completely invalid, we
silently went ahead and skipped the module load. The intention was to
only fallback when no module was found though.

Those commands now error out as expected in that case.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>