sdk/java: constructors (#9523)
* sdk/java: implement constructor
* fix javadoc fields

ensure we always set a string as the javadoc, even if empty, and ensure
we don't experience null pointer exception when looking for the documentation.
This was possible for constructors as there's no `@Function` annotation and
the code was expecting the annotation.

* ensure there's only one single @Module annotation
* ensure there's only one constructor per module
* ensure only the main object can have a constructor
* remove un-necessary dagger.json
* reverse order to read description

In case `@Module`, `@Object` or `@Function` annotation defines
a `description` field, uses it instead of the JavaDoc comment.

* raise an error if more than 1 non empty constructor exists
* register constructor with an empty name

---------

Signed-off-by: Yves Brissaud <gh@lgtd.io>