mitigate upstream otel goroutine leak in telemetry flush (#9700)
There's a problem upstream where a canceled context during
`batchSpanProcessor.ForceFlush` leaks a goroutine. I saw this actually
happen while testing stuff and it's extra bad because it leaks quite a
substantial amount of memory that's indirectly referenced in our various
span processors.

For now, this just mitigates it. I have an upstream fix that we will
hopefully pick up someday after it's released and we've upgraded.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>