logs: allow missing resources / trace id / span ID
It's valid for these to not be defined. These logs might not end up
where the user wants them as a result, but there's no need to error so
loudly about it.

We'll at least log a warning if there's no trace or span ID, but not
having a resource schema URL seems OK / normal.

Signed-off-by: Alex Suraci <alex@dagger.io>