fix(sdk/java): code completion on VS Code (#9707)
The VS Code java plugin is missing some features for years: the ability
to code complete based on the target/generated-sources files.
This allows to explicitly instruct maven/vs code by adding the generated
source directories as source directories so that code completion can
happen.

Also added some comments to the pom.xml file to help understand which
are the critical parts of the file that shouldn't be touched by the
user.

Signed-off-by: Yves Brissaud <gh@lgtd.io>