dagui: align span status with activity representation
Previously nested linked test spans would have an incrementing duration,
yet have a 'completed' status. Now the status will be running so long as
the duration is increasing. Arguably this implementation was simply out
of date.

An alternative fix would be change effect.IsRunning() to
effect.IsRunningOrEffectsRunning(), but just using the Activity is much
cheaper.

Signed-off-by: Alex Suraci <alex@dagger.io>