tui: look up calls by output digest, too (#9739)
* tui: look up calls by output digest, too

Currently for uploaded local dirs the UI shows the raw digest or a
placeholder instead of the Host.directory call that created it. This is
because the server's Host.directory actually returns a call with a
custom content digest of the blob, rather than the call digest of the
Host.directory call.

Now when querying the DB for a call we'll also look for spans that had
the digest as an output.

Signed-off-by: Alex Suraci <alex@dagger.io>

* add golden test for directory args

Signed-off-by: Alex Suraci <alex@dagger.io>

* flush only when running new command

otherwise when the command exceeds screen height it just gets flushed to
the scrollback leaving the prompt at the top of an empty screen

Signed-off-by: Alex Suraci <alex@dagger.io>

* dagui: propagate canceled status, like failed

Signed-off-by: Alex Suraci <alex@dagger.io>

* dagui: align span status with activity representation

Previously nested linked test spans would have an incrementing duration,
yet have a 'completed' status. Now the status will be running so long as
the duration is increasing. Arguably this implementation was simply out
of date.

An alternative fix would be change effect.IsRunning() to
effect.IsRunningOrEffectsRunning(), but just using the Activity is much
cheaper.

Signed-off-by: Alex Suraci <alex@dagger.io>

* Simplify Simplify

Tried this before but it didn't work out for some reason. Taking another
swing at it now that I'm not on a time crunch.

This fixes a test hang (Call/ArgTypes/SSH_Private_GitLab) and builds on
the prior work to track creators in CreatorSpans.

Signed-off-by: Alex Suraci <alex@dagger.io>

* bump oteltest

originally to fix TestCall being 'succeeded', but the real fix ended up
being proper cancel status propagation, since in practice these spans
didn't end anyway

Signed-off-by: Alex Suraci <alex@dagger.io>

---------

Signed-off-by: Alex Suraci <alex@dagger.io>