feat: refactor dagop usage
This allows returning `File`s and `Directory`s more easily, using a
wrapped immutable cache ref - this avoids needing our nasty hack to
import from the host file system: we can just write directly to our
cache ref.

To get this working with `Container.asTarball` though, we needed a
little reworking, so that's it's possible to get a reasonable filename
for the tarball. Potentially this doesn't work if the filename is
*completely* unknown until running the function? So, the
`DirectoryDagOp` becomes a generic `FSDagOp` able to handle both
`Directory`s and `File`s.

Signed-off-by: Justin Chadwell <me@jedevc.com>