replace impurity with mix of custom cache keys + field spec settings (#9719)
* replace impurity with mix of custom cache keys + field spec settings

This is a mix of smallish cleanups/preparation for more changes:

Impure is gone now and replaced with:
* `CachePerCall` - a custom cache key impl that lets the API always run
  when called but be cached after the result of that call is passed
  around (i.e. for snapshots of mutating state)
* `DoNotCache` - a setting that exists purely on the Field now rather
  than being tied into IDs. We still want to straight-up skip caching
  values for various reasons (correctness and memory usage), but there's
  no longer any need for that to be part of the ID. This replaces it.

Related to the above, we no longer cache the results of selecting
"trivial fields" (i.e. selecting a field already set on a parent object,
as opposed to actually executing a an API impl for the field). We skip
it via `DoNotCache`.
* The reason for this is that it has O(n^2) memory usage for nested
  struct cases. Selecting a series of nested fields like
  `Obj.field1.field2.field3` results in progressively caching each
  object in the chain even though you already have everything cached you
  need in the "outer" one.
* This started mattering a lot more in my trials at cross-session
  caching, so want to get it fixed now in prep for that.

A few other related things are here too, i.e. removing code referenced
to `Meta` as a field/id configuration, which hasn't been used for a
while.

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

* address pr feedback

Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Erik Sipsma <erik@sipsma.dev>