fix: don't set SSH_ASKPASS to echo
I don't really have context for why this was added, but it's definitely
not right, resulting in *always* setting a completely incorrect auth
token.

See the following:

    $ GIT_TERMINAL_PROMPT=0 SSH_ASKPASS=echo git credential fill <<EOF
    protocol=https
    host=github.com
    EOF

    protocol=https
    host=github.com
    username=Username for 'https://github.com':
    password=Password for 'https://Username%20for%20%27https%3A%2F%2Fgithub.com%27%3A%20@github.com':

This outputs the prompts into the username/password fields, instead of
*correctly* erroring out, and not attaching this as an authentication
token.

I'm not actually sure if this was causing user-facing bugs, but I hit
this while working on a total refactor of the git implementation.

Signed-off-by: Justin Chadwell <me@jedevc.com>