chore(sdk/java): improve versioning (#9767)
* chore(sdk/java): set version of java sdk

Update all the pom.xml files when calling `sdk | java | bump`
This is in preparation to release the dependencies to maven central.

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* chore(sdk/java): fix version in user module

Build the module dependencies using the version of the engine
and use that version in the user module pom.xml.

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* fix(sdk/java): fix replacement of placeholder

the `dagger-module` placeholder was in conflict with a real
`dagger-module` entry (generated sources) so make it more clear
one is a placeholder to be sure to edit the right one

Signed-off-by: Yves Brissaud <gh@lgtd.io>

* update integration tests

Signed-off-by: Yves Brissaud <gh@lgtd.io>

---------

Signed-off-by: Yves Brissaud <gh@lgtd.io>