ci: Run all Engine & CLI tests against a dev engine only
We spoke about this with Erik, Justin & Connor a few times in the last
few weeks, and the consensus is that we want all Engine tests to run
against **dev** Engines (the next version that ships), and there is
marginal value in running them against an already released version.

The thinking goes that if all tests continue running the same (or
better) in the next version of Dagger, then we are confident in the
upcoming release.

While we might want to do the same for all other jobs in this workflow,
we want to take a smaller step since we are planning to try out a
different CI runner type.

Any jobs which were flaking - egine-test-publish & scan-engine - are now
using Single Tenant (`st`) instances so there is no chance of one job
affecting the other.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>